requirements.txt
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
certifi==2024.6.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
langfuse==2.36.2
openai==1.35.3
packaging==23.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
wrapt==1.16.0
